fix endpoint url problem, add basic agent tests
