+++ tests/agents/test_kbase_agent.py
@@ -2,17 +2,6 @@ from langchain_core.language_models.llms import LLM
 from narrative_llm_agent.agents.kbase_agent import KBaseAgent
 import pytest
 
-class MockLLM(LLM):
-    def _call():
-        pass
-
-    def _llm_type():
-        pass
-
-@pytest.fixture
-def mock_llm():
-    return MockLLM()
-
 def test_kbase_agent(mock_llm):
     fake_token = "foo"
     agent = KBaseAgent(fake_token, mock_llm)
